requirements.txt
azure-storage-blob
flask